Refactor character dashboard hook into modular sub-hooks
